requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
scipy>=1.11
pyyaml>=6.0
joblib>=1.3
matplotlib>=3.8
streamlit>=1.36
gdown
